requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
requests==2.22.0
web3==5.31.3
